Pull out attachment rendering and uploads for localization to share
