+++ src/components/flow/actions/sendmsg/SendMsgForm.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/explicit-member-accessibility */
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
 import { react as bindCallbacks } from 'auto-bind';
-import axios from 'axios';
+import { AxiosResponse } from 'axios';
 import Dialog, { ButtonSet, Tab } from 'components/dialog/Dialog';
 import { hasErrors, renderIssues } from 'components/flow/actions/helpers';
 import {
@@ -15,11 +15,10 @@ import { hasUseableTranslation } from 'components/form/assetselector/helpers';
 import CheckboxElement from 'components/form/checkbox/CheckboxElement';
 import MultiChoiceInput from 'components/form/multichoice/MultiChoice';
 import SelectElement, { SelectOption } from 'components/form/select/SelectElement';
-import TextInputElement, { TextInputStyle } from 'components/form/textinput/TextInputElement';
+import TextInputElement from 'components/form/textinput/TextInputElement';
 import TypeList from 'components/nodeeditor/TypeList';
-import Pill from 'components/pill/Pill';
 import { fakePropType } from 'config/ConfigProvider';
-import { fetchAsset, getCookie } from 'external';
+import { fetchAsset } from 'external';
 import { Template, TemplateTranslation } from 'flowTypes';
 import mutate from 'immutability-helper';
 import * as React from 'react';
@@ -33,7 +32,7 @@ import {
   FormEntry
 } from 'store/nodeEditor';
 import { MaxOfTenItems, Required, shouldRequireIf, validate } from 'store/validators';
-import { createUUID, range } from 'utils';
+import { range } from 'utils';
 
 import styles from './SendMsgForm.module.scss';
 import { hasFeature } from 'config/typeConfigs';
@@ -41,30 +40,7 @@ import { FeatureFilter } from 'config/interfaces';
 
 import i18n from 'config/i18n';
 import { Trans } from 'react-i18next';
-import { TembaSelectStyle } from 'temba/TembaSelect';
-
-const MAX_ATTACHMENTS = 3;
-
-const TYPE_OPTIONS: SelectOption[] = [
-  { value: 'image', name: i18n.t('forms.image_url', 'Image URL') },
-  { value: 'audio', name: i18n.t('forms.audio_url', 'Audio URL') },
-  { value: 'video', name: i18n.t('forms.video_url', 'Video URL') },
-  { value: 'application', name: i18n.t('forms.pdf_url', 'PDF Document URL') }
-];
-
-const NEW_TYPE_OPTIONS = TYPE_OPTIONS.concat([
-  { value: 'upload', name: i18n.t('forms.upload_attachment', 'Upload Attachment') }
-]);
-
-const getAttachmentTypeOption = (type: string): SelectOption => {
-  return TYPE_OPTIONS.find((option: SelectOption) => option.value === type);
-};
-
-export interface Attachment {
-  type: string;
-  url: string;
-  uploaded?: boolean;
-}
+import { Attachment, renderAttachments } from './attachments';
 
 export interface SendMsgFormState extends FormState {
   message: StringEntry;
@@ -183,14 +159,6 @@ export default class SendMsgForm extends React.Component<ActionFormProps, SendMs
     }
   }
 
-  public handleAttachmentRemoved(index: number): void {
-    // we found a match, merge us in
-    const updated: any = mutate(this.state.attachments, {
-      $splice: [[index, 1]]
-    });
-    this.setState({ attachments: updated });
-  }
-
   private getButtons(): ButtonSet {
     return {
       primary: { name: i18n.t('buttons.ok', 'Ok'), onClick: this.handleSave },
@@ -199,183 +167,6 @@ export default class SendMsgForm extends React.Component<ActionFormProps, SendMs
         onClick: () => this.props.onClose(true)
       }
     };
-  }
-
-  private renderUpload(index: number, attachment: Attachment): JSX.Element {
-    return (
-      <div
-        className={styles.url_attachment}
-        key={index > -1 ? 'url_attachment_' + index : createUUID()}
-      >
-        <div className={styles.type_choice}>
-          <SelectElement
-            key={'attachment_type_' + index}
-            name={i18n.t('forms.type', 'Type')}
-            style={TembaSelectStyle.small}
-            entry={{
-              value: { name: attachment.type }
-            }}
-            options={TYPE_OPTIONS}
-            disabled={true}
-          />
-        </div>
-        <div className={styles.url}>
-          <span className={styles.upload}>
-            <Pill
-              icon="fe-download"
-              text="Download"
-              large={true}
-              onClick={() => {
-                window.open(attachment.url, '_blank');
-              }}
-            />
-            <div className={styles.remove_upload}>
-              <Pill
-                icon="fe-x"
-                text="Remove"
-                large={true}
-                onClick={() => {
-                  this.handleAttachmentRemoved(index);
-                }}
-              />
-            </div>
-          </span>
-        </div>
-      </div>
-    );
-  }
-
-  private handleUploadFile(files: FileList): void {
-    let attachments: any = this.state.attachments;
-
-    // if we have a csrf in our cookie, pass it along as a header
-    const csrf = getCookie('csrftoken');
-    const headers: any = csrf ? { 'X-CSRFToken': csrf } : {};
-
-    // mark us as ajax
-    headers['X-Requested-With'] = 'XMLHttpRequest';
-
-    const data = new FormData();
-    data.append('file', files[0]);
-    axios
-      .post(this.context.config.endpoints.attachments, data, { headers })
-      .then(response => {
-        attachments = mutate(attachments, {
-          $push: [{ type: response.data.type, url: response.data.url, uploaded: true }]
-        });
-        this.setState({ attachments });
-      })
-      .catch(error => {
-        console.log(error);
-      });
-  }
-
-  private renderAttachment(index: number, attachment: Attachment): JSX.Element {
-    let attachments: any = this.state.attachments;
-    return (
-      <div
-        className={styles.url_attachment}
-        key={index > -1 ? 'url_attachment_' + index : createUUID()}
-      >
-        <div className={styles.type_choice}>
-          <SelectElement
-            key={'attachment_type_' + index}
-            style={TembaSelectStyle.small}
-            name={i18n.t('forms.type_options', 'Type Options')}
-            placeholder={i18n.t('forms.add_attachment', 'Add Attachment')}
-            entry={{
-              value: index > -1 ? getAttachmentTypeOption(attachment.type) : null
-            }}
-            onChange={(option: any) => {
-              if (option.value === 'upload') {
-                window.setTimeout(() => {
-                  this.filePicker.click();
-                }, 200);
-              } else {
-                if (index === -1) {
-                  attachments = mutate(attachments, {
-                    $push: [{ type: option.value, url: '' }]
-                  });
-                } else {
-                  attachments = mutate(attachments, {
-                    [index]: {
-                      $set: { type: option.value, url: attachment.url }
-                    }
-                  });
-                }
-                this.setState({ attachments });
-              }
-            }}
-            options={index > -1 ? TYPE_OPTIONS : NEW_TYPE_OPTIONS}
-          />
-        </div>
-        {index > -1 ? (
-          <>
-            <div className={styles.url}>
-              <TextInputElement
-                placeholder="URL"
-                name={i18n.t('forms.url', 'URL')}
-                style={TextInputStyle.small}
-                onChange={(value: string) => {
-                  attachments = mutate(attachments, {
-                    [index]: { $set: { type: attachment.type, url: value } }
-                  });
-                  this.setState({ attachments });
-                }}
-                entry={{ value: attachment.url }}
-                autocomplete={true}
-              />
-            </div>
-            <div className={styles.remove}>
-              <Pill
-                icon="fe-x"
-                text=" Remove"
-                large={true}
-                onClick={() => {
-                  this.handleAttachmentRemoved(index);
-                }}
-              />
-            </div>
-          </>
-        ) : null}
-      </div>
-    );
-  }
-
-  private renderAttachments(): JSX.Element {
-    const attachments = this.state.attachments.map((attachment, index: number) =>
-      attachment.uploaded
-        ? this.renderUpload(index, attachment)
-        : this.renderAttachment(index, attachment)
-    );
-
-    const emptyOption =
-      this.state.attachments.length < MAX_ATTACHMENTS
-        ? this.renderAttachment(-1, { url: '', type: '' })
-        : null;
-    return (
-      <>
-        <p>
-          {i18n.t(
-            'forms.send_msg_summary',
-            'Add an attachment to each message. The attachment can be a file you upload or a dynamic URL using expressions and variables from your Flow.',
-            { count: MAX_ATTACHMENTS }
-          )}
-        </p>
-        {attachments}
-        {emptyOption}
-        <input
-          style={{
-            display: 'none'
-          }}
-          ref={ele => {
-            this.filePicker = ele;
-          }}
-          type="file"
-          onChange={e => this.handleUploadFile(e.target.files)}
-        />
-      </>
-    );
   }
 
   private handleTemplateChanged(selected: any[]): void {
@@ -493,6 +284,39 @@ export default class SendMsgForm extends React.Component<ActionFormProps, SendMs
     );
   }
 
+  private handleAttachmentUploaded(response: AxiosResponse) {
+    console.log('uploaded', response);
+
+    const attachments: any = mutate(this.state.attachments, {
+      $push: [{ type: response.data.type, url: response.data.url, uploaded: true }]
+    });
+    this.setState({ attachments });
+  }
+
+  private handleAttachmentChanged(index: number, type: string, url: string) {
+    let attachments: any = this.state.attachments;
+    if (index === -1) {
+      attachments = mutate(attachments, {
+        $push: [{ type, url }]
+      });
+    } else {
+      attachments = mutate(attachments, {
+        [index]: {
+          $set: { type, url }
+        }
+      });
+    }
+
+    this.setState({ attachments });
+  }
+
+  private handleAttachmentRemoved(index: number) {
+    const attachments: any = mutate(this.state.attachments, {
+      $splice: [[index, 1]]
+    });
+    this.setState({ attachments });
+  }
+
   public render(): JSX.Element {
     const typeConfig = this.props.typeConfig;
 
@@ -524,7 +348,13 @@ export default class SendMsgForm extends React.Component<ActionFormProps, SendMs
 
     const attachments: Tab = {
       name: i18n.t('forms.attachments', 'Attachments'),
-      body: this.renderAttachments(),
+      body: renderAttachments(
+        this.context.config.endpoints.attachments,
+        this.state.attachments,
+        this.handleAttachmentUploaded,
+        this.handleAttachmentChanged,
+        this.handleAttachmentRemoved
+      ),
       checked: this.state.attachments.length > 0
     };
